So, I have a few projects going at one time in this my workspace :-\ I'm going commit them separately to make it simpler to sort them out later.
This set of changes is working on a common network setup webui.

While working on this I realized a bit of the code would be common to the telnet
interface I was working on for a different project .... so I started the a common network
setting lib... it will be in another commit.