add a Makfile to so a buildbot can run tests on every checkin
Why am I adding this..?

Because, it is good to know if your change just broke the build for another project area and it's a pain to compile all of them.

Yes, it's a tiny project but, I also want to know if the could "should" compile before I try :-)

Also, it lets the public see that it DOES compile so if they are having issues then the issue is with there setup.